requirements.txt
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.3.0
langchain-text-splitters>=0.3.0
langgraph>=0.2.0
python-dotenv>=1.0.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
rank-bm25>=0.2.2
jieba>=0.42.1
numpy>=1.24.0